+++ litebox_broker_transport_linux_userland/src/lib.rs
@@ -21,7 +21,7 @@
 
 mod setup;
 
-pub mod shared_memory;
+pub mod memfd;
 
 mod unix_io;
